Bugbounty workflow moved
