+++ workflows/bugbounty/target.py
@@ -1,0 +1,12 @@
+from dataclasses import dataclass, field
+from typing import List
+
+@dataclass
+class BugBountyTarget:
+    """Bug bounty target information"""
+    domain: str
+    scope: List[str] = field(default_factory=list)
+    out_of_scope: List[str] = field(default_factory=list)
+    program_type: str = "web"  # web, api, mobile, iot
+    priority_vulns: List[str] = field(default_factory=lambda: ["rce", "sqli", "xss", "idor", "ssrf"])
+    bounty_range: str = "unknown"
